Add search page
This commit adds a search page that will be used to display the results of a search.

The search page will be rendered when the user navigates to /search/[id] where [id] is the id of the search.

The search page will display the results of the search.

The search page will also display a loading indicator if the search is still pending.